baremetal: Make libbaremetal an OBJECT library
Some toolchains rely on libc in other system libraries such as libgcc.
Linking libbaremetal as an archive would cause the (GNU) linker to
discard objects which are not needed by objects preceding libbaremetal
in the link order. As libgcc is linked at a later stage we will get a
linker error. This may be avoided by making libbaremetal an OBJECT
library and linking its objects explicitly.

Note that another solution is to set the `--whole-archive` linker flag
before `libbaremetal.a` in the linker command but there does not seem
to be an easy way to do this in CMake.